ModifyMocha: add DepCtrl; get rid of global variables; consolidations; minor optimizations
